+++ ui/podcasts.js
@@ -252,13 +252,17 @@ var podcasts = function() {
 			var target = $('#podcast_'+channel);
 			var uri = "includes/podcasts.php?populate=1&loadchannel="+channel;
 			var term = $('[name="podsearcher_'+channel+'"]').val();
+			var configvisible = target.find('.podconfigpanel').is(':visible');
 			if (typeof term !== 'undefined' && term != '') {
 				uri += '&searchterm='+encodeURIComponent(term);
 			}
 			$('i[name="podcast_'+channel+'"]').makeSpinner();
 			target.load(uri, function() {
+				if (configvisible) {
+					target.find('.podconfigpanel').show();
+				}
 				target.removeClass('loaded').addClass('loaded');
-				updatePodcastDropdown(channel,  null);
+				updatePodcastDropdown(channel, null);
 			});
 		},
 
@@ -407,9 +411,9 @@ var podcasts = function() {
 		},
 		
 		search: function() {
-		    doSomethingUseful('cocksausage', language.gettext("label_searching"));
-			var term = $('#podcastsearch').val();
 			$('#podcast_search').empty();
+		    doSomethingUseful('podcast_search', language.gettext("label_searching"));
+			var term = $('#podcastsearch').val();
 		    $.ajax( {
 		        type: "GET",
 		        url: "includes/podcasts.php",
@@ -420,12 +424,11 @@ var podcasts = function() {
 		            $("#podcast_search").html(data);
 		            $('#podcast_search').prepend('<div class="menuitem containerbox padright brick_wide sensiblebox"><div class="configtitle textcentre expand"><b>Search Results for &quot;'+term+'&quot;</b></div><i class="clickable clickicon podicon icon-cancel-circled removepodsearch fixed"></i></div>');
 		            $("#podcast_search .fridge").tipTip({delay: 500, edgeOffset: 8});
-					$('#spinner_cocksausage').remove();
 					layoutProcessor.postAlbumActions($('#podcast_search'));
 		        },
 		        error: function(data, status, thing) {
 		            infobar.notify(infobar.ERROR, "Search Failed : "+data.responseText);
-		            $('#spinner_cocksausage').remove();
+		            $('#spinner_podcast_search').remove();
 		        }
 		    } );
 		},
@@ -468,6 +471,16 @@ var podcasts = function() {
 		storePlaybackProgress: function(track) {
 			podcastRequest({setprogress: track.progress, track: encodeURIComponent(track.uri)}, null);
 		},
+		
+		globalAction: function(thing, el) {
+			el.makeSpinner();
+			var options = new Object;
+			options[thing] = 1;
+			podcastRequest(options, function() {
+				el.stopSpinner();
+				podcasts.reloadList()
+			});
+		}
 
 	}